refactor: 좋아요 기능 더티체킹으로 변경 #3
