--- src/main/java/com/wanted/socialMediaIntegratedFeed/domain/content/ContentRepository.java
@@ -1,6 +1,0 @@
-package com.wanted.socialMediaIntegratedFeed.domain.content;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ContentRepository extends JpaRepository<Content, Long> {
-}
