+++ src/main/java/com/wanted/socialMediaIntegratedFeed/domain/member/MemberRepository.java
@@ -1,0 +1,4 @@
+package com.wanted.socialMediaIntegratedFeed.domain.member;
+
+public interface MemberRepository {
+}
