+++ src/main/java/com/wanted/socialMediaIntegratedFeed/domain/post/PostRepository.java
@@ -1,0 +1,6 @@
+package com.wanted.socialMediaIntegratedFeed.domain.post;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
