--- src/main/java/com/wanted/socialMediaIntegratedFeed/domain/user/UserRepository.java
@@ -1,4 +1,0 @@
-package com.wanted.socialMediaIntegratedFeed.domain.user;
-
-public interface UserRepository {
-}
