+++ src/main/java/com/wanted/socialMediaIntegratedFeed/global/config/SecurityConfig.java
@@ -1,16 +1,11 @@
 package com.wanted.socialMediaIntegratedFeed.global.config;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.web.SecurityFilterChain;
 
+@Configuration
 public class SecurityConfig {
-
-//    @Bean
-//    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception{
-//        return (web) -> web.ignoring().anyRequest();
-//    }
 
     @Bean
     public WebSecurityCustomizer webSecurityCustomizer() {
